Rewrite the protocol almost entirely to a new model that actually looks somewhat favourable:
Each component/block is a class with property-slots for each kind of property it supports. The property-slot keeps track of a property type and a relational constraint. The relational constraint describes whether the property may have multiple values, is required, or is tied to other properties. The value in a property-slot is either a list of, or a single instance of the class denoted by that slot's property type. Each property is in turn modelled as a class with a value slot and slots for each kind of parameter it supports. Parameters, finally, are modelled as direct-slot-definition classes, that store the parameter's type constraint as default initargs.